🚚 Create Router for API auth
